fix: Ensure treesitter parse tree is up-to-date (#69)
Whenever get_location() is called, we request the treesitter parser
to refresh the parse tree for the current buffer up to date.

This may be not needed when the treesitter-highlight module is enabled,
but when the module is disabled the plugin may not display the
correct context information (as reported in #69).

Such calls will be throttled properly: refreshing the treesitter
parse tree will happen only when the buffer has changed.